Moved Coords and Grid to the Game namespace
