+++ src/Domain/Game/Field.php
@@ -3,7 +3,6 @@
 namespace Wecamp\FlyingLiqourice\Domain\Game;
 
 use Assert\Assertion;
-use Wecamp\FlyingLiqourice\Domain\Coords;
 
 class Field
 {
