+++ src/Domain/Game/Fields.php
@@ -3,7 +3,6 @@
 namespace Wecamp\FlyingLiqourice\Domain\Game;
 
 use Assert\Assertion;
-use Wecamp\FlyingLiqourice\Domain\Coords;
 
 class Fields implements \IteratorAggregate
 {
