+++ src/Domain/Game/FireResult.php
@@ -1,11 +1,8 @@
 <?php
 
-
 namespace Wecamp\FlyingLiqourice\Domain\Game;
 
-
 use Assert\Assertion;
-use Wecamp\FlyingLiqourice\Domain\Coords;
 
 class FireResult
 {
